requirements.txt
PyQt4
py2exe_py2
smartcard
pyautogui